[Fixes #162] Add rollback function on all handlers
