+++ importer/handlers/csv/handler.py
@@ -12,6 +12,7 @@ from dynamic_models.models import ModelSchema
 from importer.handlers.common.vector import BaseVectorFileHandler
 from importer.handlers.utils import GEOM_TYPE_MAPPING
 from django.db.models import Q
+from importer.utils import ImporterRequestAction as ira
 
 logger = logging.getLogger(__name__)
 
@@ -35,6 +36,10 @@ class CSVFileHandler(BaseVectorFileHandler):
             "importer.copy_geonode_data_table",
             "importer.publish_resource",
             "importer.copy_geonode_resource",
+        ),
+        ira.ROLLBACK.value: (
+            "start_rollback",
+            "importer.rollback",
         ),
     }
